catalog-backend: split out some common DB operations
